requirements.txt
numpy>=1.21.0
matplotlib>=3.4.0
seaborn>=0.11.0
Pillow>=8.0.0
torch>=1.9.0
torchvision>=0.10.0
scikit-learn>=0.24.0
pandas>=1.3.0
--extra-index-url https://download.pytorch.org/whl/cu126
tqdm
safetensors
timm
transformers
huggingface_hub
easydict
wandb
catalyst
omegaconf
opencv-python
scikit-image
opencv-contrib-python
tensorflow-datasets 
